backend-python/requirements.txt
# Core web / api
fastapi
uvicorn[standard]

# LangChain + embeddings + connectors
langchain
chromadb
tiktoken
langchain-groq            # groq integration (if you use the langchain-groq wrapper)
langchain-community       # community integrations
langchain-openai          # optional if you call OpenAI (keeps imports clean)

# HuggingFace / SentenceTransformers embeddings
sentence-transformers
transformers              # useful for some hf models

# Specific langchain provider adapters (some projects use these names)
langchain-chroma         # adapter utilities for chroma (some envs)
langchain-huggingface    # huggingface adapter bindings (if used)

# Document processing / OCR / PDF
unstructured[all-docs]   # recommended for robust PDF parsing (includes many extras)
pdfplumber               # alternate/supplemental PDF parsing
pytesseract              # system Tesseract wrapper (needs Tesseract installed on OS)
Pillow                   # imaging

# Image summarization / requests
requests

# Utilities
python-dotenv
lxml
chromadb                 # duplicate allowed, pip dedupes automatically
